# Conflicts:
#	src/main/java/de/vandalismdevelopment/vandalism/Vandalism.java
#	src/main/java/de/vandalismdevelopment/vandalism/gui/ingame/HUDElement.java

+++ src/main/java/de/vandalismdevelopment/vandalism/gui/imgui/impl/menu/impl/CustomHUDConfigImGuiMenu.java
@@ -13,7 +13,7 @@ import java.awt.*;
 public class CustomHUDConfigImGuiMenu extends ImGuiMenu {
 
     private boolean mouseDown;
-    private double lastMouseX, lastMouseY;
+    private int lastMouseX, lastMouseY;
 
     public CustomHUDConfigImGuiMenu() {
         super("Custom HUD Config");
@@ -23,21 +23,26 @@ public class CustomHUDConfigImGuiMenu extends ImGuiMenu {
     @Override
     public void render(final DrawContext context, final int mouseX, final int mouseY, final float delta) {
         final CustomHUDRenderer customHUDRenderer = Vandalism.getInstance().getCustomHUDRenderer();
+
         if (ImGui.begin("Custom HUD Config##customhudconfig", Vandalism.getInstance().getImGuiHandler().getImGuiRenderer().getGlobalWindowFlags())) {
             if (ImGui.button("Close Custom HUD Config##closecustomhudconfig")) {
                 this.setState(false);
             }
+
             if (ImGui.button("Reset Custom HUD Config##resetcustomhudconfig")) {
                 for (final HUDElement hudElement : Vandalism.getInstance().getCustomHUDRenderer().getHudElements()) {
                     hudElement.reset();
                 }
+
                 try {
                     Vandalism.getInstance().getConfigManager().getCustomHUDConfig().save();
                 } catch (final Exception e) {
                     Vandalism.getInstance().getLogger().error("Failed to save custom hud config.", e);
                 }
             }
+
             ImGui.separator();
+
             for (final HUDElement hudElement : Vandalism.getInstance().getCustomHUDRenderer().getHudElements()) {
                 if (ImGui.treeNodeEx(hudElement.getName() + "##" + hudElement.getName() + "customhudconfig")) {
                     if (ImGui.button("Reset##reset" + hudElement.getName() + "customhudconfig")) {
@@ -48,17 +53,21 @@ public class CustomHUDConfigImGuiMenu extends ImGuiMenu {
                             Vandalism.getInstance().getLogger().error("Failed to save custom hud config.", e);
                         }
                     }
+
                     ImGui.spacing();
                     hudElement.renderValues();
                     ImGui.treePop();
                 }
             }
+
             ImGui.separator();
             ImGui.spacing();
             ImGui.end();
         }
+
         final Window window = this.window();
         final double scaledWidth = window.getScaledWidth(), scaledHeight = window.getScaledHeight();
+
         for (final HUDElement hudElement : customHUDRenderer.getHudElements()) {
             hudElement.render(
                     this.mouseDown,
@@ -72,10 +81,12 @@ public class CustomHUDConfigImGuiMenu extends ImGuiMenu {
                     delta
             );
         }
+
         context.drawHorizontalLine(0, (int) scaledWidth, (int) (scaledHeight * 0.66), Color.green.getRGB());
         context.drawHorizontalLine(0, (int) scaledWidth, (int) (scaledHeight * 0.33), Color.green.getRGB());
         context.drawVerticalLine((int) (scaledWidth * 0.66), 0, (int) scaledHeight, Color.green.getRGB());
         context.drawVerticalLine((int) (scaledWidth * 0.33), 0, (int) scaledHeight, Color.green.getRGB());
+
         this.lastMouseX = mouseX;
         this.lastMouseY = mouseY;
     }
@@ -84,14 +95,17 @@ public class CustomHUDConfigImGuiMenu extends ImGuiMenu {
     public void mouseClick(final double mouseX, final double mouseY, final int button, final boolean release) {
         if (button == 0) {
             this.mouseDown = !release;
+
             if (release) {
                 boolean save = false;
+
                 for (final HUDElement hudElement : Vandalism.getInstance().getCustomHUDRenderer().getHudElements()) {
                     if (hudElement.shouldSave) {
                         hudElement.shouldSave = false;
                         save = true;
                     }
                 }
+
                 if (save) {
                     try {
                         Vandalism.getInstance().getConfigManager().getCustomHUDConfig().save();
